requirements.txt

alpaca-py>=0.30.0
binance>=0.3
elegantrl>=0.3.3 # install with: pip install --no-deps elegantrl
exchange_calendars>=3.3
gym>=0.26.0
joblib>=1.1.0
matplotlib>=3.4.3
numpy>=1.22.2
optuna>=2.10.0
pandas>=1.3.4
python-dotenv>=0.19.0
psutil>=5.8.0
pygame>=2.5.0
python_binance>=1.0.15
pytz==2021.3
scikit_learn>=1.2.0
scipy>=1.8.0
seaborn>=0.11.2
statsmodels>=0.13.2
stockstats>=0.3.2
TA_Lib>=0.4.21
torch>=1.9.1
tqdm>=4.63.0
yfinance>=0.1.63
